Added some tests for PatchBuilder replaceInsert
With the current situation of the b3_0 sources, it looks like there are no tests for the replaceInsert functionality of the patch builder. In a recent project I found out, that the behavior of the patch builder does differ if the node does exist (replace) vs. if does not yet exist (insert: will put the whole fragment instead of the selectPath). This behavior is for the application developer a bit inconvenient, since you have to check first wether the node does exist or not, and then to decide how you structure the fragment node in question.

NOTE: Therefore this test case highlights the current problem and the second assertion (L. 357, "please make me not nested") will fail.