DOMHandle: Reuse DocumentBuilder and LsParser
Creating a new DocumentBuilder and LsParser each time
DOMHandle#receiveContent gets called is very time consuming.
Fix this by using a per thread cached DocumentBuilder in
DomHandle. LSParser is created the first time receiveContent is called
and then reused.

Extracted CachedInstancePerThreadSupplier to be able to reuse it in
DocumentBuilderFactories.

Fixes #1054.

Signed-off-by: Wagner Michael <maffelbaffel@posteo.de>