JAVA-225: Bumped OkHttp to 4.10.0
Also bumped slf4j and logback, as those have almost no transitive dependencies. Bumped the Jackson dependences in functional-tests as well for consistency with the rest of the project. 

Also bumped the kotlin-stdlib version used by ml-development-tools to 1.6.20 to match the kotlin version used by OkHttp 4.10.0. 

Removed some unnecessary Maven repos. 

Finally, bumped up the Gradle wrapper to the latest 6.x version. This is in preparation for eventually going to the latest 7.x; just want to ensure all is well first with the latest 6.x version.