Added pom.xml file to try out github's Dependency Graph
I'd like to get this added to the master branch so that I can see what Insight / Dependency Graph shows in github - in particular, I want to see what the Dependents view shows. Adding it to the develop branch won't help because github only cares about what's on the master branch. And this has to be a Maven pom.xml file instead of a Gradle build.gradle file since the Dependency Graph feature only supports Maven and not Gradle.