JAVA-246: Added docCols(String[])
This is necessary for when the user only wants a subset of descriptor columns, but doesn't need to map them - having to specify a map just for this use case would be very annoying. 

Also added DocumentWriteOperationImpl(uri, metadata, content) and added newWriteOp to get rid of a lot of duplication in tests.