Extracted CONTRIBUTING.md info
Didn't want this buried in the README. It's a little awkward still because .github/CONTRIBUTING.md exists with a bunch of boilerplate on how to create and submit a PR. That can be cleaned up later. The key is for the top-level CONTRIBUTING.md to describe how to actually develop and test an enhancement. 

Also removed the "What's new" content from the README. This was a bit random - some for Java 4, some for Java 5.3. The CHANGELOG has all this data too, so I removed it from the README.