#1036: Deprecated unused mimetypes argument in ResourcesServices
Was able to remove this completely in OkHttpServices, as the argument wasn't used for anything. But can't remove it from ResourceServices since that's a public interface. So deprecating the argument in the 6 methods that have it. 

Adjusted each test and example not to pass in a mimetype argument anymore; those were misleading since it implied that the argument was having some impact.