Splitting functional tests into separate Gradle tasks
The hope is that when these run in the regression, the fast ones will all succeed and won't be affected by the results of the slow ones, which will run after the fast ones. 

Also added another call to mlDeploy before running the functional tests to ensure that the expected test app is present.