Added securityContextType property to test plumbing
This removes a lot of hardcoding we have that forced the usage of digest authentication. Also removed some hardcoding of assumed port numbers. 

Intent is to make it easier to add support for running all the tests against a reverse proxy, which will most likely require basic authentication. 

Added fast functional tests to the PR pipeline - figuring this will be good to leave in for now as a quick way of ensuring a good chunk of the functional tests still work.