DEVEXP-148: Extracted AuthenticationConfigurer to simplify auth code
Only doing this for basic/digest as we have sufficient tests for both of those authentication types. Intent is to both make this part of OkHttpServices more readable and to make it easier to add token-based auth (i.e. via a separate class instead of stuffing it into OkHttpServices). 

I did refactor some of the kerberos/cert/saml code, as all 5 auth branches were needlessly doing null checks on getSSLContext and getTrustManager. `configureSocketFactory` is already checking to see if those are null or not before doing anything with them.

Added `getTrustManager()` to `SecurityContext` to simplify accessing this too - the code for accessing this was needlessly complicated given that all implementations of `SecurityContext` already had this public method.