Refactoring fast functional tests to create client via one method
The intent is to make it simple to add a basePath in a follow up PR. No functionality change otherwise, just funneling every DatabaseClient construction via one method. 

"Slow" functional tests aren't impacted yet as it's not likely that we'll test those yet via the reverse proxy server. That's because they're hitting a variety of ports - which they don't need to do, but each needs to be refactored first.