DEVEXP-144: DatabaseClient now supports basePath
The app changes were fairly simple here, with the most notable part being the addition of HttpUrlBuilder to encapsulate the logic for building a base URL, which has to account for basePath now. Added a bunch of unit tests for this too. 

Most of the rest of this PR is test stuff, specifically the new support for setting up a reverse proxy server and configuring the tests to use that server instead of talking to ML directly.