DEVEXP-143: Extracted methods for constructing an OkHttp Client
This is a pure refactor, no functionality change. The intent is to make it easy for the upcoming MarkLogic cloud authenticator to reuse this code for making an HTTPS call to get a token. 

Added a basic SSL test to ensure that SSL connections can be made. This makes use of JUnit 5 extension support so that we can easily add more tests in the future that require SSL.