DEVEXP-149: Fixing buggy usage of StringUtils
Moved the global dependencies to the precise places where they're needed to minimize confusion. Removed usage of StringUtils in DocDescriptorUtil, which was passing in the tests but failed in the real world since commons.lang3 is a compileOnly dependency.

A PR coming soon will move all the cookbook examples to a separate project, allowing us to get rid of several of these compileOnly dependencies to further avoid the chance of this bug occurring again.