DEVEXP-145: Moved examples to subproject
distZip still includes the necessary files, it just now pulls them from the separate subproject. 

Was able to remove some exclusions from the jar and sourcesJar targets, as those were excluding examples-specific files that are no longer in marklogic-client-api. 

Removed dependencies from marklogic-client-api that were only needed for cookbook examples. 

Otherwise, there are no changes here to any of the example programs, nor to the tests that exercise them. This is simply moving everything to a new subproject and then adjusting the Gradle files.