Added test for onBatchFailure
This test now produces a real failure that results in the onBatchFailure listener being invoked. 

We had about a dozen tests calling onBatchFailure before, but none of them were actually doing something to cause the failure listener to be invoked. Tellingly, removing the `onBatchFailure listener` used in each of these tests still resulted in the tests passing, which confirms that the listeners were never being invoked. So the onBatchFailure call was removed from each of these tests to prevent anyone from thinking it was actually being tested in those tests.