Builder now inits SSL context when protocol and trust manager exist
Ran into this while testing, where I was providing both a protocol and a "trust all" trust manager, but the builder failed to create a client. In this scenario, the builder is able to initialize the SSLContext itself since it has a trust manager from the user, so it now does so.