Added custom mappings to reverse proxy server
This supports testing ml-gradle where a user wants to both deploy a REST app server and load REST extensions, which requires loading them via that REST app server. The user will need to define `restBasePath` . The RPS now makes it easy to define a mapping from the value of `restBasePath` to whatever port the REST app server is listening on.