Removed factory methods that have been broken since 4.2.0
These don't work because they don't provide a security context. And since the 4.2.0 release and a change to OkHttpServices on 2019-01-07 that made a security context required, they haven't worked. I was going to deprecate them initially, but in the event that a pre 4.2.0 user upgrades directly to 6.1.0, I think a compile time error would be far better than waiting for either method to fail at runtime (and both are guaranteed to fail).