DEVEXP-469 Optimistic locking failures now throw concrete exceptions
Users no longer need to parse out exception messages. And now `TestOptimisticLocking` doesn't need to either, which allowed me to delete a bunch of code from it in favor of asserting on specific exception classes. 

This is a non-breaking change as the new exceptions are subclasses of `FailedRequestException` and contain the same error message.