DEVEXP-562 Can now prevent Accept-Encoding=gzip header being used
DatabaseClientFactory also supports multiple configurators now. This was essential for testing this new story, but it also aligns with what a user would expect a method named "add" to do (as opposed to "set"). `clearConfigurators` was also added both for testing purposes and so a user could remove all the configurators in case one or more should no longer be used.