ml-development-tools now uses test-app
Got rid of the custom code for setting up / tearing down databases and app servers. Changed the test code to use port 8013, as java-client-api is already using 8012 and 8014. 

Updated gitignore as well so you can run the tests and not have 1400+ changes sitting in your branch. 

The 3 Java files that were updated are all test files and were updated when I ran `generateTests`.