DEVEXP-563 Added geohashSubhashes to Optic plan builder
This is all the code-generated stuff for DEVEXP-563. Will use a second PR for the manually-added stuff. 

Most of this is just copyright updates that were made by the code generator. 

Note that the code generator also updated PlanGeneratorTest. We have some manual mods in there so that some tests only run on MarkLogic 11, so I had to ensure those were retained.