Bumping Gradle wrapper to version 8
This will make supporting Java 21 easier, and also includes a few updates based on deprecated Gradle stuff. 

Also removed the "dist" stuff from marklogic-client-api. AFAICT, we have not updated nor published this zip in a long time. It seems intended for someone to download and have a local copy of the examples and javadocs. It causes issues with Gradle 8, so removing since we haven't used this in a long time.