Bumping slf4j-api to 2.0.16
This may be the root cause of the logging issue I'm running into with Flux - at least according to https://www.slf4j.org/codes.html#StaticLoggerBinder , as I verified that slf4j-api 1.x is sneaking onto the classpath. This should otherwise be a no-op for all of the Java Client, it doesn't care what version of slf4j-api it's using.