MLE-17146 Added support for annTopK
I removed the restriction in BaseTypeImpl that requires every value to be a subclass of `BaseArgImpl`. That puts a big burden on the server in that it must accept "wrapped" values for primitive values. The server function `annTopK` does not support "wrapped" values for `k` or `queryTolerance`, and thus that restriction caused things to break.

Not sure we'll keep this change yet.