Finished switching over to ManageClient
No longer seeing any leaked resources when running the "functional" tests, which is most likely to getting rid of the manual use of OkHttpClient.

Deleted the disabled TestBug18993 class as well. I think it's testing something that is OBE now, which pertains to spaces in the URI of a document. We have other tests for that verifying that it works in MarkLogic 12.