MLE-21420 Refactor: Consolidated TransformerFactory construction
No effect on behavior here. This is just to help with an issue in the Spark connector, where the Saxon-HE TransformerFactory is being used. Having all TransformerFactory.newInstance() calls in one place makes it easier to think about how it's used in the Java Client.

Also good from a security perspective - i.e. all uses of TF are being configured with the necessary precautions.