MLE-24504 Bumped Jakarta and removed javax.ws
The javax.ws.rs-api dependency was only needed for a collection class that wasn't needed.

Also forcing the usage of commons-lang3 as Black Duck mysteriously thinks an instance of 3.7 is being brought in as a direct dependency somehow.

Also switched to latest nightly build for local testing, as Jenkins is using that too.